requirements.txt
pandas
python-dotenv
pydantic
matplotlib
d-geolib